Sync flow with Android
