+++ Internal/Debug App/Sources/View Controllers/MerchantHeadlessCheckoutAvailablePaymentMethodsViewController.swift
@@ -105,6 +105,14 @@ class MerchantHeadlessCheckoutAvailablePaymentMethodsViewController: UIViewContr
         }
     }
     
+    @IBAction func onVaultManagerButtonTap(_ sender: Any) {
+        let vc = MerchantHeadlesVaultManagerViewController.instantiate(settings: settings,
+                                                                  clientSession: clientSession,
+                                                                  clientToken: clientToken)
+        self.navigationController?.pushViewController(vc, animated: true)
+    }
+    
+    
     // MARK: - HELPERS
     
     private func showLoadingOverlay() {
